# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CommonPlace API — smart essentials platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List auto-refill subscriptions for current user
      responses:
        "200":
          description: Subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create an auto-refill subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubscriptionBody"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    put:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update a subscription (skip, modify, cancel)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubscriptionBody"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: cancelSubscription
      tags: [subscriptions]
      summary: Cancel a subscription
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Admin ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth ────────────────────────────────────────────────────────
...
    Subscription:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        interval:
          type: string
          enum: [2weeks, 1month, 2months, 3months]
        mode:
          type: string
          enum: [auto, confirm]
        nextDeliveryDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending_payment, active, paused, cancelled]
        stripeSubscriptionId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, productId, productName, quantity, interval, mode, nextDeliveryDate, status, createdAt]
    CreateSubscriptionBody:
      type: object
      properties:
        productId:
          type: integer
          format: int32
          minimum: 1
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
        interval:
          type: string
          enum: [2weeks, 1month, 2months, 3months]
        mode:
          type: string
          enum: [auto, confirm]
        notifyDaysBefore:
          type: integer
          format: int32
          minimum: 0
          maximum: 30
      required: [productId, quantity, interval, mode]
    UpdateSubscriptionBody:
      type: object
      description: >
        Fields for updating a subscription. `quantity`, `interval`, and `mode`
        are billable terms and may only be changed when the subscription is in
        `pending_payment` status. Attempting to change these fields on an
        `active` or `paused` subscription will return a 422 error. To change
        billable terms on an active or paused subscription, cancel it and create
        a new one.
      properties:
        quantity:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          nullable: true
        interval:
          type: string
          enum: [2weeks, 1month, 2months, 3months]
          nullable: true
        mode:
          type: string
          enum: [auto, confirm]
          nullable: true
        status:
          type: string
          enum: [active, paused, cancelled]
          nullable: true

    # ── Admin ────────────────────────────────────────────────────────
...